refactor: refactor DDD for CV
