+++ src/CV/Domain/Repository/ProfessionalExperienceRepository.php
@@ -1,38 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\CV\Domain\Repository;
 
 use App\CV\Domain\ProfessionalExperience;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
 
-/**
- * @extends ServiceEntityRepository<ProfessionalExperience>
- */
-class ProfessionalExperienceRepository extends ServiceEntityRepository
+interface ProfessionalExperienceRepository
 {
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, ProfessionalExperience::class);
-    }
+    public function save(ProfessionalExperience $professionalExperience): void;
 
     /**
-     * @return list<ProfessionalExperience>
+     * @return array<ProfessionalExperience>
      */
-    public function findAll(): array
-    {
-        $qb = $this->createQueryBuilder('professional_experience')
-            ->orderBy('professional_experience.startDate', 'DESC');
+    public function findAll(): array;
 
-        return $qb->getQuery()->getResult();
-    }
-
-    public function findOneLatest(): ProfessionalExperience|null
-    {
-        $qb = $this->createQueryBuilder('professional_experience')
-            ->orderBy('professional_experience.startDate', 'DESC')
-            ->setMaxResults(1);
-
-        return $qb->getQuery()->getOneOrNullResult();
-    }
+    public function findOneLatest(): ProfessionalExperience|null;
 }
