+++ src/CV/Domain/Route.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\CV\Domain;
+
+enum Route: string
+{
+    case INDEX = 'app.cv.index';
+}
