+++ src/CV/Infrastructure/Doctrine/Repository/ProjectORMRepository.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\CV\Infrastructure\Doctrine\Repository;
+
+use App\CV\Domain\Project;
+use App\CV\Domain\Repository\ProjectRepository;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @extends ServiceEntityRepository<Project>
+ */
+class ProjectORMRepository extends ServiceEntityRepository implements ProjectRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Project::class);
+    }
+
+    public function save(Project $project): void
+    {
+        $this->getEntityManager()->persist($project);
+    }
+
+    public function findAll(): array
+    {
+        $qb = $this->createQueryBuilder('project')
+            ->orderBy('project.date', 'DESC');
+
+        return $qb->getQuery()->getResult();
+    }
+
+    public function findOneLatest(): Project|null
+    {
+        $qb = $this->createQueryBuilder('project')
+            ->orderBy('project.date', 'DESC')
+            ->setMaxResults(1);
+
+        return $qb->getQuery()->getOneOrNullResult();
+    }
+}
